+++ src/components/Instructions/Instructions.js
@@ -2,13 +2,16 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import * as actions from "../../actions";
 import requireAuth from "../requireAuth";
+import InstructionsViewer from "./InstructionsViewer";
+import ProductChooser from "./ProductChooser";
 import "./InstructionsStyle.scss";
 
 class Instructions extends Component {
   render() {
     return (
       <div className="instructions-page">
-        <p>Hi there!</p>
+        <ProductChooser />
+        <InstructionsViewer />
       </div>
     );
   }