+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.15.0] - 2021-03-05
+
+### 🔥 Added
+
+- (WIDS) created full Instructions component / page
+- (WIDS) created ProductChooser component
+- (WIDS) created InstructionsViewer component
+- (WIDS) created LineProductChooser component
+- (WIDS) created PartNumberProductChooser component
+- (WIDS) created StickerScanProductChooser component
+- (WIDS) created ZkScanProductChooser component
+- (WIDS) created LinkBlock component
+- created Loader component
+
 ## [1.14.0] - 2021-03-02
 
 ### 🔥 Added